refactor(db): extract db-search crate for dense search + embedding (Phases 5+6 of #796)
Moves dense_search.rs and embedding.rs out of db into a new db-search
crate, along with EmbeddingClientConfig. Addresses #792 design question
with the "standalone leaf crate" answer.

## What's extracted

- crates/db/src/dense_search.rs -> crates/db-search/src/dense_search.rs (543 LOC)
- crates/db/src/embedding.rs -> crates/db-search/src/embedding.rs (344 LOC)
- EmbeddingClientConfig moved from db::database to db-search::config

## Why this works cleanly (unlike Phases 1-4)

Unlike the fetchers and patch module, dense_search and embedding:

- Are already free functions, no impl<S: Store> DB<S> blocks to unwind
- Take &EmbeddingClientConfig (plain config struct) and &E: QueryExecutor
  (external trait) as parameters
- Touch zero DB<S> private state
- Have exactly one external consumer (defra-node)
- Use only query, schema, document, reqwest as deps

db-search depends on query/schema/document with no dependency back to db,
so db depends on db-search cleanly. EmbeddingClientConfig comes from
db-search; db re-exports it via pub use db_search::EmbeddingClientConfig.

## Dense vs SE coupling

Dense search and searchable-encryption (SE) tag generation are largely
independent today. SE lives in db-merge (extracted in PR #735) and
handles HMAC-SHA256 tag generation for encrypted indices. Dense search
handles vector similarity for RAG-style retrieval. They share nothing
except being "non-trivial search features" — no shared types, no shared
code paths. Keeping them in separate crates is the right call; merging
them into a single db-search-and-encrypted crate would be forced
unification.

## Internal callers

Internal db callers rewrite imports:
- crate::embedding::set_embedding -> db_search::set_embedding
  (5 sites in crates/db/src/auto_commit_mutator/)
- crate::dense_search items are re-exported via `pub use db_search as dense_search`
  so the rest of db works unchanged

## Test plan

- cargo check --workspace clean
- cargo test -p db -p db-search all passing
- cargo clippy --workspace --all-targets -- -D warnings clean
- cargo fmt --all applied

## Diff

2 files moved (~887 LOC), 1 new crate skeleton, 1 new config module,
5 import rewrites in auto_commit_mutator, 1 re-export consolidation in
db/src/lib.rs.

Closes #792. Related: #796 (epic).
