+++ Cargo.toml
@@ -20,6 +20,7 @@ members = [
     "crates/db-blocks",
     "crates/db-nac",
     "crates/db-merge",
+    "crates/db-search",
     "crates/keyring",
     "crates/identity",
     "crates/zanzibar",
