+++ crates/db-search/Cargo.toml
@@ -1,0 +1,35 @@
+[package]
+name = "db-search"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Vector search, embedding client, and hybrid BM25+dense retrieval for DefraDB"
+
+[dependencies]
+# Internal crates
+query = { path = "../query" }
+schema = { path = "../schema" }
+document = { path = "../document" }
+
+# HTTP client (for embedding generation API calls)
+reqwest = { version = "0.12", features = ["json"] }
+
+# Serialization
+serde.workspace = true
+serde_json.workspace = true
+
+# Error handling
+anyhow.workspace = true
+thiserror.workspace = true
+
+# Logging
+tracing.workspace = true
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
+
+[lints]
+workspace = true
