+++ crates/db-search/src/dense_search.rs
