+++ crates/db-search/src/embedding.rs
