+++ crates/db/src/auto_commit_mutator/update.rs
@@ -15,7 +15,7 @@ impl<S: Store + 'static> AutoCommitMutator<S> {
         let mut modified_fields = modified_fields;
         let embedding_config = self.db.options().embedding_config();
 
-        let generated = crate::embedding::set_embedding(
+        let generated = db_search::set_embedding(
             &collection.schema().vector_embeddings,
             &mut doc,
             false,
